moved PIT-model to corresponding PIT example.
